Make the disable function safe for non-Ubuntu OSs
Specifically /usr/sbin/invoke-rc.d does not
exist under Centos 6.4

+++ stemcell_builder/lib/helpers.sh
@@ -13,8 +13,11 @@ function codename() {
 }
 
 function disable {
-  mv $1 $1.back
-  ln -s /bin/true $1
+  if [ -e $1 ]
+  then
+    mv $1 $1.back
+    ln -s /bin/true $1
+  fi
 }
 
 function enable {
